MNT: Verified metaclass magic
While this now seems sort of stupid, it was almost done by the
time I figured out it was stupid to do this, so no we have a
way to valiate that super_state_machine is working as promised.
Even though it was working as promised to begin with. Uh, so,
yeah. Major victory today.