MNT: address minor thread-related paranoia
We have lots of threads running around, if `self.detectors` is looking
at a global object and that is changed from another thread this can lead
to inconsistent behavior (reading something we did not trigger etc).

Grab a copy of the detectors to be used outside the loop and then work
on that.