ENH: first pass at adding logbook integration
Add a 'logbook' command which allows the scan to ship a message
string + kwargs to the run engine.  The current run_start_id
is added to the kwargs (overwriting anything that is there).

The string + the dict are then passed to the self.logbook
callable as `self.logbook(logbook_msg, kwargs)`.