ENH: add wait_for message
This message takes a list of futures as the object.

These are passed as-is to `asyncio.wait` which blocks
the progress of the main co-routine until all of the Events
have been set.

This is the same mechanism used by the block groups, but exposed
to the user is a rawer manner.