ENH: add delay in resuming
@cmazzoli As requested.

Subtly changed logic it `__call__`, now use the thread safe
call_soon to kick off a call_later to set the Event.  Drop the
local reference to the event as soon as we send it off (but before
it actually gets set).  This should allow multiple suspend -> resume ->
suspend -> resume cycles to behave reasonably.