FIX: logic to mark run open to end of function
If the insertion in to mds fails then we reach the
finally block of `_run` without inserting a run_start
document, but with the run marked as open.  This causes
inserting the run_stop to fail an raise which means the
`loop.stop()` never gets called and the scan hangs.

The issue of exceptions in the finally block will be
addressed later.