FIX: fix a bug in resume logic
When stopping the event loop in asyncio, the currently running co
routine runs until it hits the next yield.  By putting a sleep at
the top of the main run loop we force the `_run` co-routine to yield
control before trying to get the next message to process.

If this is not here, it would get the next message to process, but
suspend just before processing it so that a) that message would be in
the msg_cache (at the end) and b) that message will be processed
_before_ the first message in the new generator added to the generator
stack by `resume`.