FIX: on timeout mark asyncio.Event as done
This is needed so that if the run engine is blocking on a device
to complete motion it will break out of that to come back to the top of
the `_run` loop to see the stashed error.

We do not want to `cancel` the main task on an time out to allow plans a
chance to re-try.