API: hold onto control of signal handling
In SignalHandler do not relinquish control of the signal handling
until the context manager exits.  This will ensure that our clean up
code in the `finally` block will run all the way through but makes tha
code essentially un-interuptable.  This is a feature, not a bug, as that
code takes care of ensuring a number of assumptions about the state of
the run enginge and the hardware are enforced.