API/FIX: clear msg cache on un-cacheable msg
This fixes a subtle bug where there are some messages which are
not (always) safe to replay due to being part of a context pair.

If a checkpoint has been requested before these messages are seen
they reset the msg cache.  This is effectively an implicit `checkpoint`
immediately after the message if the user supplied a checkpoint and
is a no-op if the user has not supplied a checkpoint (or has issued a
clear_checkpoint).