ENH: handle corner cases
If there is a plan is the stack that wants to do some clean up via
messages that got to the RunEngine, but the top plan does not (ex, if
there has been a plan with rewinding or a suspender plan have been
inserted on top.  Thus, if the same exception comes right back
out (meaning the plan did not try to yield something), try throwing the
exception at the next plan.