FIX: close down all plans in the case of abort
Abort goes through a different code path than `stop`.  In `stop` and
we stash and exception which gets run through the loop (and passed down
to the plans).  In `abort` (and failed pausing) we call `cancel` on the
task which causes a CancelledError to be raise someplace in the
co-routine (`_run`).  This by-passes code to shove the exception down
into the plans and just exits the _run loop.

Thus, in the `finally` explicitly close all of the plan still in the
plan stack.