TST: cover case of badly behaving plan
If a plan (which does not get fully executed due to an earlier plan
in the stack raising an exception (internally, not via injection
from the RE) which aborts without trying to re-handle the exception.

This tests that if one of the plans is a badly behaved generator the
warning is correctly printed.