MNT: handle switching state to 'running' carefully
Move the idle -> running transition into `_run` to prevent the case
where the RE would be in a running state, but the code between where it
was flipped and where the `_run` coro was actually started raised.

This left the RE in a state where it was 'running', but had to task,
thus no part of the API (which relies on the finally block of the coro
to flip the RE back to idle) would be able to return the RE to an idle state.