MNT: simplify plan.throw exception handling
There is no need to check if the same exception bounced back out
or not.  If a plan has raised then that generator is dead so pop it off
the stack and if there are any more plans left, try throwing the current
exception down.

This has the limitation that we are loosing traceback context (but no
worse that it was before).