FIX: allow `None` to be a valid group for waiting
This makes the code work the way that @danielballan thought it
worked.

All objects that are set/trigger/kickoff/collect will have their status
objects hooked up to a an `asyncio.Event` and if the Status object
reports that it failed will cause an exception to be raised.

Before this, if there was not a non-None group, the RunEngine would not
check the success or failure of the action.