FIX: do not keep old exceptions around
We are never actually creating extra tasks so this is useless and the
traceback object is keeping the failed task alive so we see it every
time.

We are creating handles, not Tasks, elsewhere in the code.

If we do start to create other tasks, we must keep track of them to
clean them up rather than relying on a global weak-ref dict in the
Task class.