MNT: relax jsonschema validation of json "array" type
Accept list or tuple for 'array'.  This will not ensure that documents
type-for-type round trip, but means we can have things at the ohpyd
layer that return tuples (for the immutability) and not have to have
extraneous list calls through out the RE.   So far all of our
consumers will take lists and tuples equally.