DOC: work around bug-fix in ophyd
Signal now no longer lies about the type always being 'number',
however the initial point in the example plan was 15 which caused
the first call to `describe` on the motor to report that the
'motor' data key was of type 'integer'.  Future data points, which are
floats, would then cause LiveTable to raise.

Not worried about fixing this in LiveTable as we expect type-stability
of the read values.