FIX/TST: Update RunEngine._create and tests for event-model 1.6.0.
This should address the issue that has broken the bluesky CI since 29
August.

The "name" (i.e. stream name) item in Event Descriptors was introudced
in bluesky 0.4.x in February 2016. In the latest release of event-model,
1.6.0, we ensure that _if_ "name" is present its value must be a string.
This requirement breaks some old tests in bluesky, which generated Event
Descriptors with ``{'name': None, ...}``. We still allow name to be
absent, so that old data will still validate successfully.

It is not expected that much (or maybe any) production data has this
problem, because getting ``None`` requires manually spelling out
``Msg('create')``. The default name from ``bluesky.plan_stubs.create``
or high-level plans that call it is ``'primary'``.