TST: tweak timeout on test
There are now too many polling loops and we are getting increased
latency from buffer bloat.

In this test, the event is timing out concurrently with the second
sigint being processed and (at least locally) we are losing the race
between the plan exiting cleanly and the pause request getting
triggered.