MNT: do not try to cancel task on stop or abort
Calling cancel would break out of a plan that was waiting in a long
sleep or motion, there is a race between the RequestAbort and
RequestStop exception being dutifully thrown down the plans and the
CancelledException being raised.  This results in violations of the
`_run` loop invariants.

Operationally, this is a moot-point because stop and abort are
typically called from a 'paused' state (and
`request_pause( deferred=False)` still does cancel the task to break
out of long waits and return control promptly).