FIX: process any outstanding signals
If a callback emits a signal via the telepotor in the Qt-aware
callbacks there is a race condition where the run will finish and
return before the function on the slot has had a chance to run.  This
means that not all of the callbacks will be fully done Qt main loop is
stopped.  Adding the `processEvents` will clear any event on the queue
when it is called (but not handle any subsequent events added in the
process of processing those events) so this is guaranteed to have
finite run-time.