MNT: be gentler about running the Qt event loop
Create an EventLoop to run rather than running the main app.  This
re-enters the event loop on the current thread (creating one if needed?).

This a bit of future proofing for when we want to use this in an
always running QApplication (as we don't want to quit the actual
application!).