TST: start the SIGINT thread timers in the plan
This avoids a rare race condition on travis when it takes longer to
set up the coverage / pytest introspection that we have delay on the
sigint timer resulting in the sigints going to pytest (instead of us!)
and pre-maturely exiting the test suite.

Avoid his by starting the timers inside the plan which is definitely
after have had a chance to install our signal handler.