API: put exit status on control exceptions
We had been setting state on RE, then raising these exceptions in
the plans, and then relying on the state in the RE to over-ride the
exit status on the close_run yielded by the plans.

This now injects the exit status into the exceptions which is picked
up by the run_decorator and yielded out correctly.

We still need to track the state on the RE so we get the right exit
status for plans that do not handle the exceptions on their own.