API: add run entry to Msg namedtuple
This is an API change, but this is is the most reliable way to make
sure we can inject the run into any message.  If we tried to inject it
into args or kwargs there is a chance of collision with their current
usage (as they are passed through to the underlying objects in several
cases).