FIX: make sure the teleporter lives on the main thread
Previously, we were creating a new teleporter for each callback which
seemed like a good idea from not having too much shared state.
However, this means that we ended up creating the teleporter on the
background thread which then teleported the callback execution back to
the background thread which is not useful.

Instead of returning the teleporter class we now return a live
object.  The signature on signal now passes the (name, doc, object of
interest).  The one registered slot then calls the object as
`obj(name, doc, escape=True)` which will then run the actual callback
on the thread that the teleporter was created on (which is ideally the
main thread).