FIX: add draw_idle to LiveGrid and LiveScatter
Previously we were relying on the qt kicker to trigger the re-draws,
move the logic into the callbacks instead.

This is both more robust (it is still safe for something external to
check the stale status and trigger e re-draw) and more-correct (the
callback object knows when it is back to a consistent state and should
be re-drawn).