FIX: Harden the zmq RemoteDispatcher
If we got malformed messages into .zmq.RemoteDispatcher we would fail
which would kill the loop pulling from the socket.  However, because we
started the loop in a task and then did `loop.run_forever` the task would be
dead, holding on to its exception, and the main ascyio loop would keep running
resulting in a "dead" process with no error.

This adds an option to be "strict" and raise if bogus payloads come in or
to drop the bad messages on the floor and continue.

Either way, we will get messages on the client.