API/MNT: do not pass loop kwarg if we do not have to
The loop parameter on many asyncio functions was deprecated in py38 and
removed in py310.

The solution we had implemented was to pass None on py38 and above, however in
py310 the parameter is removed and passing `loop=None` is a TypeError.

This switches to unpacking a dictionary (which is empty on py38+)

This also removes an used loop parameter and attribute on RunBundler.  This was
a vestigial attribute from when the RunBundler logic was pulled out of the
RunEnigne and was never used.