FIX: handle case when flyer has no configuration
It is documented as optional so it should be optional.  Previously the RE call
would run OK, but because we had no configuration, there was no configuration
key entry for it.  This PR ensures that there is at least an empty dictionary
as we do not need to convey in the documents between "the device reported
configuration and it was empty" from "the device did not report any
configuration".