FIX: Also clear the describe cache on configure
This is to fix an issue at FXI where they were changing the number of
images (and hence the shape) on an AD.  Through enough stage/unstaging the
Resource documents had correct information, but because we were caching the
result of `describe` the first time we saw the object all of the streams were
erroneously reporting that they had the same shape which was eventually causing
tiled to error out.

For this to work it makes the assumption that any PV values that might sneak
into the descriptor (rather than the configuration) would also be configuration
values (if they are not, the configure message will fail on the call to the
configure ophyd method).

This is also slightly over-kill for many cases where nothing in describe
changes, but we have no other channel to clear these caches.