MNT: rip out tracking handles from thread-safe callbacks in asyncio
The way the bridging between ophyd objects <-> RE works for anything with a
Status object:

 - ophyd returns a status object
 - RE subscribes to the done callback on the status object
 - that callback closes over an asyncio.Event
 - RE also holds onto the asyncio.Event for use in _wait
 - because the done callback runs on a different thread (the epics thread not
   the RE thread) we use call_soon_threadsafe which schedules a sync functions
   to be called by the EventLoop as soon as possible
 - call_soon_threadsafe returns an asyncio.Handle object which _only_ lets
   you cancel or ask if the callback was canceled (these are what we
   were caching and calling "tasks").  You can not even ask if it is done or
   started, it can only cancel.

It is not clear to me if calling this a task is due to a change in CPython (we
started when asyncio was a bit provisional) or due to us not understanding
things when we wrote this code (but I would bet on door number 2).  The idea
was that if we needed to cancel things we would want to cancel these tasks,
however:

 - the Handle is created by a callback only when that status object is done
 - from inside the callback there is no way to get back to the Handle we
   are holding
 - There should be no harm in letting these run

I tried changing `status_tasks` to be a dict mapping p_event ->
Handle, however that opened up a race condition where the
callback gets run before Handle can be added to the dictionary
which lead to race-conditions.

Caching the Handles without clearing them is harmless for short plans, however
every `set`, `trigger`, and `kickoff` adds a Handle to the cache.  For
long (days to months) long plans this will eventually become a problem.

In conclusion:

 - These Handles are very limited and transient
 - Capturing them robustely is very hard
 - The callbacks should be very fast so the Handle is only useful very
   transiently
 - I do not think there is harm is always letting these callbacks run.  The
   only possible failure path I see is that it is possible we may clobber
   an exception with a failure.  If that is the case, we should work that
   out mare generally as that was already a risk but with an added side of a
  race condition.