FIX: make trigger and read careful
Previously we would get confusing errors when:

  - event had been opened
  - read or describe raise any errors once
  - a wrapper (such as the baseline pre-processor) tries to take a reading as
  part of plan clean up

The error would be that you could not open an event with an event open which
would effectively drop the original error (the one you actually need to debug
the issue!) on the floor.

This could also be fixed by improving the multi-exception handling in the RE,
but this is a simpler fix.