FIX: handle case of rewinding to before the beginning of a stream
The issue is that:

 - we track the event count (for seq_num which is non-monotonic and
   can repeat if rewound) in `Bundler._sequence_counters`
 - we snapshot the counters at checkpoints
 - on rewind we clear the _event_counters dictionary and then .update it
   from the copy

In the case of implicitly creating a stream between a checkpoint and a
pause (so there is `declare_stream` message) the above process effectively
drops our tracking in the event counters.

In `_prepare_stream` we ensure that the needed counter exists (and starts at 1)
so replicate that logic here.

Closes #1647