FIX: account for CPython's deprecation of 3 arg throw signature
We had been passing all 3 arguments through and letting Python sort it out, but
as of 3.12 the 3 argument version is deprecated so use *args, **kwargs to
blindly forward through whatever the user gave us.  This make sure that we do
not spuriously trigger the deprecation warning nor do we hide a warning the
user should get.

closes #1817